+++ lib/ex_doc.ex
@@ -27,7 +27,7 @@ defmodule ExDoc do
       ExDoc.Markdown.configure_processor(markdown_processor_options)
     end
 
-    docs = config.retriever.docs_from_dir(config.source_beam, config)
+    docs = config.retriever.docs_from_dir(config)
     find_formatter(config.formatter).run(docs, config)
   end
